feat: support union types
